Add ability to set color capability on loggocolor
I've run across a situation where I would like to force loggocolor to write colors (in this case, writing to a termcap view which implements io.Writer). However, `github.com/juju/ansiterm` does not see this as a terminal - rightly so - so this branch adds a new constructor that calls `ansiterm.SetColorCapable` for the Writer before returning it. It leaves the current API unchanged.