logger: Select loggers via wildcards
The following is very tentative exploration about if it's possible to
introduce wildcards for selecting loggers.

Normal pattern:

  a.b.x.c; a.b.y.c; a.b.z.c

Wildcard pattern:

  a.b.*.c

The problem with the approach atm is what to do when we don't find any
loggers for that pattern. At the moment it creates them, but we don't
want to create a wildcard logger, so maybe we just try and find the
parent logger to the *:

  a.b.*.c would return a.b

Additionally we're using a regexp for access, because it's simple, but
it's not performant. Although the chances of using a wildcard in
production code is low and this should be for gaining access to logs.