Logger: Label loggers
The following allows loggers to be tagged with labels, these labels can
be used when applying the configuration. With this change users of the
logging library can change the level of a series of loggers using the
tagged label.

The driver behind this is to better expose loggers with in Juju. If a
user wanted to see all http client connections, then a logger will
be tagged with `GetLogger('???', HTTP)`. Using the model-config
within Juju, it will be possible to set `[HTTP]=TRACE` and all HTTP
loggers will be set to the appropriate level.

---

Concerns:

 1. No validation on namespaces for loggers are performed, other than
 ensuring that ";" and "." are split. Meaning that "[]" for label syntax
 might already exist in the wild?
 2. What do we do about malformed "[", or "]" labels?

---

Alternative syntax for labels:

 1. "<>"
 2. "{}"