Logging: Structured logging
This basically adds structured logging to loggo. This is very tentative
exploration into better logging output for juju. Structured logging
offers better introspection and have better tooling than just using
grep/less/cat et al.

The problem we have here, is that Message in loggo.Entry is just a fat
dump of rubbish. I think the way forward here is to open up a new
function to pass anything to logging and so it can be encoded via the
writer.

Something similar to:

    logger.With("component", "client").Infof("message")

Then it can be passed internally when Infof is actually called.