feat: slog writer
To take advantage of the new logging package from the go 1.21 release,
the loggo package starts incorporate these new changes. There are
features that we could start to add to loggo to then pass directly to
slog. The migration over to slog could then be done in stages.

For now, it might be useful to at least play around with slog and see
if there are any advantages to it.