Implement attributes to the loggo package
In order to correctly interface with the log/slog package and to
take advantage of structured logging, we now introduce attributes.
Currently Info and Infof are different, in the fact that Info is
the structured variation. It expects every argument to be a attribute
after the initial message string.

This then allows us to have a slog backend without migrating away
from loggo directly. We have all the benefits of loggo plus we can
take advantage of the new logging infra.