Inherited logger labels
This introduces the ability to inherit log labels. Prior to this
patch, loggers only allowed labels for the scoped leaf node of
a logger. Any child leaf nodes did not have the ability to know the
parent label.
In a similar vein to tags, we add labels to the module. This then
allows to look up the parent on forking of the child logger. As
there is no way to edit the labels of a module it's safe to clone
those labels at forking time. This removes the need to walk the
parent stack when attempting to log a value. For deeply nested stacks
this could be quite expensive.