Remove the logger label concept
Labels should not be applied to loggers. There is no way to reference
them directly from a root module. Labels for this concept should be
more like Attributes, something that can be dynamic for a given
logger, although I suspect that they should just be sent via the
Logf method.