Customise the call depth
If we wrap the logger call in another type, then when we call logger
methods, the call depth will be off by 1. To fix this, we want to
be able to state that the call depth when constructing the logger.

This will ensure that the file names point to the right files for
consistency reasons.