Prevent module panic
If the module is empty then attempting to get the labels from it
will cause a panic. This can happen if you create a non-initialized
logger.

    var logger loggo.Logger

It only happens in tests, and the intention is to just write those
logs to /dev/null. Either way we should be backwards compatible with
the logger setup and prevent the panic.